Use OIDC trusted publishers
This stores the role to assume in a secret, which is configured as an
environment secret. While I'm unclear on whether this really needs to be
a secret (Rubygems suggests it doesn't have to be), it does have the
benefit that we can copy the workflow verbatim between various gems.